requirements.txt
pillow~=11.1.0
requests~=2.32.3
tkmacosx~=1.0.5